% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curatedHumanMarkers}
\alias{curatedHumanMarkers}
\title{Curated cell type markers for hand-selected cell types (Human)}
\format{:
  A list of gene symbols for different tissues/cell-types}
\source{
\url{http://compbio.mit.edu/ACTIONet/}
}
\usage{
curatedHumanMarkers
}
\description{
Curated cell type markers for hand-selected cell types (Human)
}
\keyword{datasets}
