% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{ENCODE}
\alias{ENCODE}
\title{Encodes an integer or integer vector into a hashid string.
All numbers must be non-negative integers.}
\usage{
ENCODE(int, settings)
}
\arguments{
\item{int}{Integer or integer vector to encode}

\item{settings}{Settings list generated by hashid_settings}
}
\value{
hashid string
}
\description{
Encodes an integer or integer vector into a hashid string.
All numbers must be non-negative integers.
}
