% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{hash}
\alias{hash}
\title{Maps an integer to a string.
Generated string will be inversely proportional to alphabet length.}
\usage{
hash(number, alphabet)
}
\arguments{
\item{number}{Integer to hash}

\item{alphabet}{Possible letters for string.}
}
\value{
hashed string
}
\description{
Maps an integer to a string.
Generated string will be inversely proportional to alphabet length.
}
