% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\docType{data}
\name{hashid_defaults}
\alias{hashid_defaults}
\alias{DEFAULT_ALPHABET}
\alias{DEFAULT_SEPS}
\alias{RATIO_SEPARATORS}
\alias{RATIO_GUARDS}
\title{Default Values for hashid settings}
\format{An object of class \code{character} of length 1.}
\source{
http://www.hashids.org
}
\usage{
DEFAULT_ALPHABET

DEFAULT_SEPS

RATIO_SEPARATORS

RATIO_GUARDS
}
\description{
Default alphabet, separators, and 
   ratio of character separators and guards for hashid
}
\keyword{datasets}
