% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{SPLIT}
\alias{SPLIT}
\title{Splits a string based on a set of splitting characters}
\usage{
SPLIT(string, splitters)
}
\arguments{
\item{string}{String to split}

\item{splitters}{set of splitting characters as a string}
}
\value{
split vector of characters
}
\description{
Splits a string based on a set of splitting characters
}
