% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{unhash}
\alias{unhash}
\title{Unhashes a string to an integer based on alphabet.}
\usage{
unhash(hashed, alphabet)
}
\arguments{
\item{hashed}{String to unhash}

\item{alphabet}{Set of letters used for hashing}
}
\value{
Unhashed integer
}
\description{
Unhashes a string to an integer based on alphabet.
}
